refactor(roulette): 拆分插件结构，迁移配置与玩家列表
- 新增 player.py，迁移 Player/PlayerList 类
- 新增 config.py 中 ShotConfig dataclass，迁移文案
- 更新注释与字段说明
